backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
motor
passlib[bcrypt]
python-jose
loguru
cryptography